# openapi.yaml
openapi: 3.1.0
info:
  title: Steady — operational API
  version: "1.0.0"
  description: >
    Steady is a Next.js App Router application. Its member-facing behaviour is
    delivered through **React Server Actions** (RPC-style, invoked from forms and
    client components — see docs/api.md for the full action inventory and their
    authorization rules), not through a conventional REST surface. Only a small
    number of **read-only JSON HTTP endpoints** exist for operational
    visibility, and they are documented here; they expose no secrets and no
    member data. The member program endpoints Handoff 10 added to the mobile
    API are documented here as well, under their own tag.
servers:
  - url: https://steady-emdr-demo.onrender.com
    description: Demo deployment
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/mobile/v1/thought-records:
    get:
      tags: [mobile-programs]
      operationId: listThoughtRecords
      summary: The member's own thought records, with the signed screen copy
      description: >
        Private to the member (Handoff 10 2B). `state` is "open" or "not_today"
        (the cautious tier, ceiling 6). Records carry the member's words by
        question, never a 0 to 10 strength. 404 while the content is not signed.
      responses:
        "200": { description: "{ state, pendingClinicalReview, copy, records }" }
        "404": { description: Not available }
    post:
      tags: [mobile-programs]
      operationId: saveThoughtRecord
      summary: Save a thought record
      description: >
        Every field runs the crisis pre-filter before anything is saved. When
        strengthBefore is 8 or more, the client offers the grounding skill
        first (copy.strong), as the web does.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [situation]
              properties:
                situation: { type: string }
                feeling: { type: string }
                strengthBefore: { type: integer, minimum: 0, maximum: 10 }
                thought: { type: string }
                supports: { type: string }
                against: { type: string }
                balanced: { type: string }
                strengthAfter: { type: integer, minimum: 0, maximum: 10 }
      responses:
        "201": { description: "{ id }" }
        "409": { description: "{ crisis: true } — nothing saved" }
        "422": { description: "{ error: not_today | write_something }" }
...